Move OPENSSL_strcasecmp() and related to o_str.c
Otherwise the implementation is unnecessarily duplicated in legacy.so.

(Merged from https://github.com/openssl/openssl/pull/18282)
