+++ include/crypto/ctype.h
@@ -79,7 +79,4 @@ int ossl_ascii_isdigit(const char inchar);
 # define ossl_isxdigit(c)       (ossl_ctype_check((c), CTYPE_MASK_xdigit))
 # define ossl_isbase64(c)       (ossl_ctype_check((c), CTYPE_MASK_base64))
 # define ossl_isasn1print(c)    (ossl_ctype_check((c), CTYPE_MASK_asn1print))
-
-int ossl_init_casecmp_int(void);
-void ossl_deinit_casecmp(void);
 #endif
