+++ include/internal/cryptlib.h
@@ -157,5 +157,6 @@ char *ossl_ipaddr_to_asc(unsigned char *p, int len);
 char *ossl_buf2hexstr_sep(const unsigned char *buf, long buflen, char sep);
 unsigned char *ossl_hexstr2buf_sep(const char *str, long *buflen,
                                    const char sep);
-
+int ossl_init_casecmp_int(void);
+void ossl_deinit_casecmp(void);
 #endif
